requirements.txt
fastapi
uvicorn[standard]
openai
python-dotenv
jinja2
motor